upstream: lcd: Add CONSOLE_SCROLL_LINES option to speed console
When the cursor position gets to the end of the LCD console we normally
scroll by one line. This adds an option to increase that value.

Console scrolling is often slow, and if a large amount of output is
being sent, increasing this option to 10 or so will speed things up
considerably.

Change-Id: Ie0512ab1b98fa1aa9ca14ef42409cb9c8bbaf00c
Reviewed-on: https://gerrit.chromium.org/gerrit/25218
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>