upstream: input: Add generic keyboard input handler
Add a module which understands converting key codes (or scan codes)
to ASCII characters. It includes FIFO support and can call back to
drivers to read new characters when its FIFO is empty.

Keycode maps are provided for un-modified, shift and ctrl keys.

The plan is to use this module where such mapping is required.

Series-changes: 3
- Add new patch for generic keyboard input handler

Series-changes: 4
- Implement key auto-repeat in the input layer

Change-Id: Ieea38f5b81300ad8e34de04589c502401041fd2a
Reviewed-on: https://gerrit.chromium.org/gerrit/20529
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>