upstream: sandbox: add concept of sandbox state
The state exists through the life of U-Boot. It can be adjusted by command
line options and perhaps later through a config file.  It is available to
U-Boot through state_...() calls (within sandbox code).

The primary purpose of this is to contain the "hardware" state.  It should
only be used by sandbox internal code.

Change-Id: Ie3a97d05cff3ca3f6df5bb350afb14926c15ebb3
Signed-off-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/23777
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>