upstream: sandbox: add getopt support
This adds simple command-line parsing to sandbox. The idea is that it
sets up the state with options provided, and this state can then be
queried later, as needed.

New flags are declared with the SB_CMDLINE_OPT_SHORT helper macro,
pointers are automatically gathered up in a special section, and
then the core code takes care of gathering them up and processing
at runtime.  This way there is no central place where we have to
store a list of flags with ifdefs.

Change-Id: Ie2a12f3ed0885686f792bf464c181d94b7bcff74
Signed-off-by: Simon Glass <sjg@chromium.org>
Signed-off-by: Mike Frysinger <vapier@gentoo.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/23779
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>