gen: Side step the tmp file when building device trees which fixes includes
The upstream version of this Makefile ran the preprocessor on the device tree
source, putting the resulting file in the build directory structure. It then
ran the device tree compiler on the temporary file from there. Because the
device tree compiler resolves includes relatively and the source files weren't
copied into the build directory, includes stopped working.

Instead of funnelling through a temp file, it attaches all the tools together
in a pipeline. Second, it explicitly cds into the directory with the source
(since there isn't a source file now) before running the compiler. Because the
source and destination paths are absolute this doesn't affect them, but now
all the includes are relative to the current working directory and dtc can
find them. The temp file is still created so that if there are errors with
line numbers relative to the preprocessor output it's easier to figure out
what that line actually says.

BUG=chromium-os:26787
TEST=Built for Stumpy

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: Ifb71f5b0e175cd359d25dde35735f388c655d617
Reviewed-on: https://gerrit.chromium.org/gerrit/16712
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@chromium.org>
Tested-by: Gabe Black (Do Not Use) <gabeblack@google.com>