Don't ever return from cold_reboot() function
cold_reboot() should wait until the reset kicks in, instead of
returning and continuing in an undefined code path. Without this
patch, vboot_twostop will return, and secure_boot() will start
printing an error message, which results in a sporadic # on
the screen when pressing space or ESC on the dev screen in order
to go to recovery mode.

BUG=chrome-os-partner:7683
TEST=boot to dev screen on lumpy, press ESC or space. Observe
     there is no # character printed before going to recovery
     mode.
Signed-off-by: Stefan Reinauer <reinauer@google.com>

Change-Id: Ic2ee570032686e48603f0fb3b1ec9cbfae9451bc
Reviewed-on: https://gerrit.chromium.org/gerrit/17007
Tested-by: Stefan Reinauer <reinauer@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Ready: Stefan Reinauer <reinauer@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/21676
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>