spl: Add parametric board initializer
Even though it is sometimes infeasible to configure the board with a
device tree, since a device tree could exceed SPL space limitation in
some cases, we still prefer a table-driven configuration mechanism such
that no recompilation is needed to update the configuration parameters,
rather than hard-coding board initialization parameters,

To do this, we create a small binary table is appended to the end of the
SPL image, which SPL reads configuration parameters from.  The table can
be updated via a binary editor; so no recompilation is needed.

Signed-off-by: Che-Liang Chiou <clchiou@chromium.org>

BUG=chromium-os:25459
TEST=Build okay
Change-Id: Ic037060d51fd22743beeb2474ffadc8be22d7621
Reviewed-on: https://gerrit.chromium.org/gerrit/17731
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>