config: Pull in the appropriate SPI configurations
The exynos-common.h file pulls in a few defines based on a check for
CONFIG_ENV_IS_IN_SPI_FLASH. We need to have this enabled before we
include the common header. Without this we wouldn't be compiling the
cmd_sf and winbond files.

BUG=None
TEST=Built
Change-Id: I272f288dd95312f47844e6646d230e6179ab21dd
Signed-off-by: Abhilash Kesavan <a.kesavan@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/17795
Commit-Ready: Simon Glass <sjg@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>