gen: Make the SCSI and USB drivers compile with the CONFIG_SYS_64BIT_LBA option
A cherry pick, format string editing was required to compile.

These drivers wouldn't compile when the CONFIG_SYS_64BIT_LBA option is turned
on because the used fixed size data types in their exported functions when they
should have used lbaint_t for the block count parameter. That meant that when
the sizes happened to be the same, when using a 28 bit LBA, the driver would
build, but when it wasn't, a 48 bit LBA, things broke.

This change adjusts the signatures to use the right type and makes small
adjustments in the affected functions.

BUG=chrome-os-partner:8180
TEST=Built for emeraldlake2 with the CONFIG_SYS_64BIT_LBA option enabled and
saw its compile errors go away.

Change-Id: I7443b5976d3d06b82070344011fba46a28bd77de
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/18058
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/21679
Reviewed-by: Vadim Bendebury <vbendeb@chromium.org>
Tested-by: Vadim Bendebury <vbendeb@chromium.org>