exynos: Add a parameter descriptor to the SPL param block
We want to be sure that a tool can safely update the parameter block of
any U-Boot SPL it is presented with. To do this, we list the parameters in
the block in a string, one character for each.

Also drop the padding at the end, and replace it with a size. The padding
does not serve a useful purpose at present, and it requires manual
maintenance. Checking a size is safer.

Signed-off-by: Simon Glass <sjg@chromium.org>
BUG=chromium-os:25459
TEST=build and boot U-Boot on Daisy
Change-Id: Ib4fe277eab93ecb97c4107f5a8f52722e1297b2d
Reviewed-on: https://gerrit.chromium.org/gerrit/18003
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>