arm: Define a hang() function for SPL u-boot.
The hang() function is needed when we use a private libgcc (USE_PRIVATE_LIBGCC).
In normal u-boot image, the hang() function is defined in arch/arm/lib/board.c,
or common/board_f.c if CONFIG_SYS_GENERIC_BOARD defined.

But for SPL u-boot, the board.o will not be included because of its large size.
As a result, it causes a compilation error of undefined reference to hang().

This fix is to define a hang() function specified for SPL u-boot.

BUG=chromium-os:25459
TEST=build with and without USE_PRIVATE_LIBGCC successfully:
$ make chromeos_daisy_config
$ make
$ make clean
$ USE_PRIVATE_LIBGCC=yes make

Change-Id: I0e08552b52a9e1b69495aab9e283119c5c8c5ca6
Signed-off-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/18227
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>