Use keyboard controller for hard reset instead of CF9
For some still unknown reason writes to localtion 0xcf9 do not cause
the Link to reboot, they cause it to shut down instead. While this
will have to be investigated and fixed, this change modifies the code
to use the keyboard controller (implemented by the EC on Link) to
restart the system.

Once the 0xcf9 problem is resolved, this change could be reverted.

BUG=chrome-os-partner:8397
TEST=manual
  . when u-boot tires restarting the system it now reboots. before
    this change it would just shut down.

Change-Id: I8076f897304f705e20ec8f2e30cce17d7fdd31c4
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/18332
Reviewed-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/21682
Reviewed-by: Stefan Reinauer <reinauer@google.com>