cros: config: Bring back non-verified boot script
These scripts are used for non-verified boot on a Chrome OS device, and
are brought back from tegra-common.h on 2011.06 branch, which was the
last place we maintained these scripts.

Although these scripts were maintained in tegra-common.h, it does not
imply that they were Tegra-specific.  Instead, they are actually quite
platform-independent and so are maintained in chromeos.h now.

Signed-off-by: Che-Liang Chiou <clchiou@chromium.org>

BUG=chromium-os:27839
TEST=Build okay

Change-Id: Ie06b0d29ab355da2952e0b208146c0035cd608f1
Reviewed-on: https://gerrit.chromium.org/gerrit/18348
Tested-by: Che-Liang Chiou <clchiou@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>