Reduce RW BIOS data sections to 512K from 976K
This saves 500ms in the RW BIOS path and still leaves 130K
overhead from the current U-boot size (372K) which should be
more than enough for any future updates on stumpy/lumpy.

This should be re-visited if/when vboot is smart enough to
just read+validate the actual binary within the section.

BUG=chrome-os-partner:8518
TEST=boot lumpy/stumpy through RW BIOS path and gather
timestamps to see 500ms improvement.

Change-Id: Ia714b4b95245f02f6e781247f820ca915bb403f5
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/18372
Reviewed-by: Stefan Reinauer <reinauer@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/21684
Tested-by: Vadim Bendebury <vbendeb@chromium.org>