exynos: Move peripheral ID into a common header and rename
We need to start setting up a common clock framework since at present
this code is spread throughout drivers, and it will only get worse.
As a first step, make use of the peripheral ID concept used by pinmux,
move it into a common header, and rename it to PERIPH_ID. We can then
start using it for clocks.

Note: SMC911X is a bit of misnomer - really it should be SMC. We will
deal with this later if needed.

BUG=chromium-os:28651
TEST=build and boot on daisy

Change-Id: I662d53a18bf5e219471db2f20600bd54e5f6a367
Reviewed-on: https://gerrit.chromium.org/gerrit/19431
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>