exynos: fdt: Add SPI bindings for exynos5250
These basic bindings will serve for now. As we have no clock definitions
in the fdt yet, we add the concept of a peripheral ID as a work-around.

This value is defined for exynos in U-Boot's periph.h file.

This requires a new dtc with symbolic support. So far this has not been
forthcoming from upstream, so I have brought in a modified version of
Stephen Warren's patch (sent several times to the device-tree-discuss
list since September 2011) in the interim.

BUG=chromium-os:29007
TEST=build and boot on Daisy
Change-Id: I184335e56b0162d5ec7ab1c1bb1dae0458f92908
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/20166