exynos: Add a function to decode a peripheral ID from the fdt
At present there is no concept of a clock in the exynos device tree.
We need access to peripheral IDs in order to adjust clocks and so on.

Add a function to convert a device tree value into a local peripheral
ID in U-Boot. For now the function is one-one.

By keeping this code in a function we can adjust the mapping or even
the binding later with minimal impact on driver code.

BUG=chromium-os:29007
TEST=build and boot on Daisy

Change-Id: I24a5b3afafbc266bd1be07253f822cda3be792a8
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/19736