exynos: Move zq_mode parameters to run-time config
Move zq_mode_dds, zq_mode_term parameters to run-time, and make
zq_mode_noterm as simple flag, which is what is intended. Note that the
shift/mask operations are done by the code, not the data.

Also add a dmc_ prefix to a dmc function that I am updating anyway.

BUG=chromium-os:28972
TEST=build and boot on daisy

Change-Id: I80b7be27521e3bc380637c146ec1bcbe6e4519a2
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/20141
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Alim Akhtar <alim.akhtar@samsung.com>