exynos: Move t_wrrdcmd to run-time config
This was called rddata_margin before, but is named T_WrRdCmd in the
rev 0.03 datasheet. Here it seems that PHY_CON0_RESET_VAL is used without
always setting up t_wrrdcmd, so we leave the value as is, and mask out
the bits when updating t_wrrdcmd in the field.

BUG=chromium-os:28972
TEST=build and boot on daisy

Change-Id: Iaa052042b6a96af7e16a9145ddda7fc438fc69a5
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/20144
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Alim Akhtar <alim.akhtar@samsung.com>