exynos: Use dmc_config_mrs() from both ddr2 and ddr3
It seems that the config_mrs() and direct_cmd() functions do roughly
the same thing. The only difference is the delays (which are
questionable anyway) and the need to send ZQINIT.

Leave the delays in the common code, and make ZQINIT a mem_timing
option. Then switch to use common code for this, and put a dmc_ prefix
on the function while we are at it.

BUG=chromium-os:28972
TEST=build and boot on daisy

Change-Id: I6c64ff8fd0840f7a7ccc7f943c3f86be5e76caec
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/20150
Reviewed-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-by: Alim Akhtar <alim.akhtar@samsung.com>