chromeos_daisy: update kernel_devnum and dev_extras
Since we want daisy on all the time on daisy, well, let's just add
it to the default u-boot. SMDK users can clear it.

Also, with the new 3.3 kernel, the device numbering for emmc vs sd has
swapped, so flip them in the default config instead of having
everybody do it themselves on all boards.

BUG=none
TEST=build for daisy, boot without doing any modifications to the environment

Change-Id: Iaf591d68668b7de9ce0b25a4f7247efe2cf21e3f
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/20408
Reviewed-by: Simon Glass <sjg@chromium.org>