x86: Provide tick counter and frequency reference for Intel core architecture
Some u-boot modules rely on availability of get_ticks() and
get_tbclk() functions, reporting a free running clock and its
frequency respectively. Traditionally these functions return number
and frequency of timer interrupts.

Intel's core architecture processors however are known to run the
rdtsc instruction at a constant rate of the so called 'Max Non Turbo
ratio' times the external clock frequency which is 100MHz. This is
just as good for the timer tick functions in question.

BUG=chromium-os:26787
TEST=manual
  . after enabling configuration necessary for this code to compile
    in, added a printout statement in main.c to report the clock
    frequency and to print a few strings at get_tbclk() tick
    intervals. The clock frequency was correctly reported at 1800 MHz
    and the printouts came out one second apart.

Change-Id: I7485159f888f836c986973b24a69dcf1bd8f9e90
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/20406