exynos5: Split GPIOs into 6 parts
GPIOs are currently split into 4 banks, but two of them have holes in them.
This means that extra code is required to deal with the two special cases.
Split the GPIOs into 6 parts, each of which has no holes.

Change the function which calculates the GPIO bank from the GPIO number to
be table-driven. This will allow it to work for different sub-architectures
with different numbers of GPIO parts.

BUG=chromium-os:30253
TEST=build and boot on daisy; build for trats, s5p_goni
See that GPIO setup by SPI port 4, and eMMC, still works.

Change-Id: I5e0ca7274d22affc70b19427b4d4f6c2f11ab447
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/21485
Reviewed-by: Alim Akhtar <alim.akhtar@samsung.com>