drop: exynos: add random delay after USB 2.0 PHY setup
In order to detect properly the USB devices connected to the USB 2.0
PHY, it seems that we need to wait for something to setup/stabilize
after putting up the PHY settings.
The current documentation is too thin to guess what is the bottleneck
here for the timings.
Let's put a semi-random value which works with my "bank" of USB devices
and revert this later if we can figure out the exact value and location for
this delay.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:8949
TEST=boot a Daisy board with various USB keys inserted and check that
they are all detected properly.

Change-Id: Ia27dbec9f5ab70c181ed38a7d632c84b11a6238f
Reviewed-on: https://gerrit.chromium.org/gerrit/20863
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Commit-Ready: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>