exynos: Make the smdk5250 board initialize the i2c busses with the fdt
The s3c24x0 driver was recently modified so that it can be configured using
the device tree. This change adds a call to the function that reads that
configuration information into the board init for the smdk5250.

BUG=chrome-os-partner:9025
TEST=Built and booted on Daisy without this change and noted what i2c devices
were found using the i2c probe command and the i2c dev command to move
between busses. Booted with this change and saw that the same devices were
found. Booted some i2c busses commented out of the device tree and saw that
those buses were no longer probe-able or selectable.

Change-Id: I4a35fae102a67380c506c7c36f02157cd1850b11
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/20913