gen: Add option to default to most compatible conf in a fit image
When booting a fit image with multiple configurations, the user either has to
specify which configuration to use explicitly, or there has to be a default
defined which is chosen automatically. This change adds an option to change
that behavior so that a configuration can be selected explicitly, or the
configuration which has the device tree that claims to be compatible with the
earliest item in U-Boot's device tree.

In other words, if U-Boot claimed to be compatible with A, B, and then C, and
the configurations claimed to be compatible with A, D and B, D and D, E, the
first configuration, A, D, would be chosen. Both the first and second
configurations match, but the first one matches a more specific entry in
U-Boot's device tree. The order in the kernel's device tree is ignored.

BUG=chrome-os-partner:9163
TEST=Built with this behavior enabled and saw U-Boot select the only kernel
in a normal, single fdt image. Used patches from Olof to build a multi-fdt
image and saw U-Boot select the correct FDT. Modified U-Boot's compatibility
list and saw the selected configuration change as expected, following the
rules above.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: I1d29eaf7d54fd2e524aec2a8b2d81e2b23bd5bee
Reviewed-on: https://gerrit.chromium.org/gerrit/21440
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>