exynos: mmc: Tidy up mmc driver configuration
We should allow the fdt to specify multiple MMC ports, and use the aliases
section to determine the ordering. Also, the concept of an 'id' in the
peripheral node is not necessary, nor commonly used - punt it.

Also move pinmux into the driver, since it has all the information needed.

BUG=chrome-os-partner:8887
TEST=build for daisy

Change-Id: I45c5af16088f6a611d8e6ebb6cb87f8cefe0f156
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/21306
Reviewed-by: Alim Akhtar <alim.akhtar@samsung.com>
Tested-by: Alim Akhtar <alim.akhtar@samsung.com>