cros: Fix boot_device to probe multiple types of device
In commit b0ba047 the boot_device was refactored but introduced a small
bug that has not been found until now.

When adding device info to an array, boot_device did not advance the
index of the array; so when it probed the next type of device, it
overwrote device that had been probed.

This bug has not been found partly because it can only be triggered when
you insert two types of removable device (or have two types of fixed
device in the machine, for that matter).

Signed-off-by: Che-Liang Chiou <clchiou@chromium.org>

BUG=chromium-os:28077
TEST=insert USB key and SD card
     run `vbexport_test init && vbexport_test diskinfo`
     see it probes both USB key and SD card correctly

Change-Id: I889cb916f3ce942ff5dbf2fda0a0fefc4520cfd0
Reviewed-on: https://gerrit.chromium.org/gerrit/21309
Reviewed-by: Che-Liang Chiou <clchiou@chromium.org>
Tested-by: Che-Liang Chiou <clchiou@chromium.org>