exynos: serial: config: Enable LCD support for smdk5250
This reverts commit 9e6173dfd2828f96d91fec881de82d80e664904a.

Re-reverting the revert made by Che-Liang. The problem was, in daisy
the LCD was not getting powered on. And since we were directing our
serial logs on the LCD(which was not getting turned on), it looked like
the u-boot did not boot up the Daisy boards.

BUG=chrome-os-partner:9310
TEST=tested on Daisy-I. The board boots up and could see the
console log on LCD

Change-Id: I2a01ef345483e28a59772cfac4daadb2d3cb9f9e
Signed-off-by: Hatim Ali <hatim.rv@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/21398
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>