exynos: mmc: Refactor MSHCI driver driver setup
We don't need two init functions with the same signature, and there is
no real reason why the config structure cannot be used both with and
without CONFIG_OF_CONTROL. Also, gd->fdt_blob is NULL when
CONFIG_OF_CONTROL is not used, so there is no need for special-case code
here.

We want the driver to support multiple peripherals, so we might
as well make that change here. Also, the pinmux should be done in the
driver, and the GPIO to enable the device set up here also.

Adjust the drivers and board file to tidy this up.

This is quite a large change and could perhaps be split. I am not sure
if that would make it better or worse.

BUG=chrome-os-partner:8887
TEST=build and boot on daisy; See that mmc still works

Change-Id: Id6620338e44d62b9a1056a728eeb9b11ae50e4c0
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/21303
Tested-by: Alim Akhtar <alim.akhtar@samsung.com>
Reviewed-by: Alim Akhtar <alim.akhtar@samsung.com>