exynos: spl: Add the function of vsprintf() and panic() in SPL u-boot
We want to enable UART in SPL u-boot. The console library is required.
But some functions we needed in the console library depends on a bunch
of library in libgeneric, like lib/ctype.o, lib/div64.o, lib/string.o,
and lib/vsprintf.o. Adding them makes the SPL u-boot too large and not
fit into the expected size.

This CL mocks these functions in SPL, like vsprintf() and panic() in
order to cut its dependency.

We won't implement all full functions of vsprintf() and just make it print
the original string instead of formatted one.

BUG=chrome-os-partner:9021
TEST=build and run u-boot successfully; detailed test pending to later CLs.

Change-Id: Ib91a21ef0589b17c29641d41a6ec9b475101160c
Signed-off-by: Tom Wai-Hong Tam <waihong@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/21750
Reviewed-by: Simon Glass <sjg@chromium.org>