x86: Remove stray printf which confuses the serial_fdt driver
There was a stray printf which duplicated an existing one slightly above it
in arch/x86/lib/board.c which was likely the result of an improper merge. The
printf happened to be after a flag was set indicating the code was relocated
but before the relocation had actually happened. Some code in serial_fdt was
called in that context, and the inconsistent state made it break.

BUG=chrome-os-partner:9022
TEST=Built and booted on Stumpy and saw some serial related problems clear up.

Change-Id: Ia5af6721e29066e5228391b072e6bfc3562e93cc
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/21754
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/21896
Commit-Ready: Gabe Black <gabeblack@chromium.org>
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>