gen: usb: Limit number of blocks per USB mass storage transaction
An earlier change (5dd95cf9) modified USB driver such that the amount
of memory handled in a single transaction is limited by 20 KBytes.
This worked well for the DVD usb controller which needed this fix (and
used 2048 byte blocks). But causes problems with the devices with
smaller block size, as the calculation yields excessive block counts.

This change limits block count by 20, which was the limit before the
aforementioned change and works with Link.

BUG=chrome-os-partner:9451
TEST=manual
  . build a new firmware image
  . install it on a Link device
  . attempt USB based recovery
  . observe the recovery attempt succeed (it was failing because of
    failing USB device reads before the change)

Change-Id: I2834a0d6763082c984af8563df7fe028f2f09d55
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/21830
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>