exynos: config: Move dsim configuration values into the config
The dsim.h header file which defines the register layout and some constants
for the exynos5 also defines constants which are used to configure the device.
Not all of these are actually constants, though, and they may need to be set
to different things in different contexts. This change attempts to move the
variable settings (as opposed to bit masks, offsets, etc.) into the smdk5250
config and out of dsim.h.

BUG=chrome-os-partner:9310
TEST=Built and booted on Daisy and, with other changes, saw output on the LCD.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: Ib02d558f08c5d4c6406cfef6282c03a110d13873
Reviewed-on: https://gerrit.chromium.org/gerrit/21856
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Ready: Gabe Black <gabeblack@chromium.org>