exynos: mshci: change the clock call sequence
MSHCI clock dvider is independent of the mshci driver operation,
These clock values should be set the the init time rather then
during operation. There are cases when only one mshci instance is active,
when kernel boots up and tried to access the other instance of which
divider is not set may fail.

BUG=none
TEST=booted daisy and saw clock divider sets properly at boot.

Change-Id: I6c55099abefa081b1f724d9165d6ed8dfbd78ea1
Signed-off-by: Alim Akhtar <alim.akhtar@samsung.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/22085
Reviewed-by: Simon Glass <sjg@chromium.org>