exynos: Add new parameter chip_to_configure
Although in exynos there are 2 chips per channel in the memory
controller, but only chip-0 is configured during initialization.
So adding a new parameter 'chip_to_configure' in the param list.
In future if we need to configure both the chips, just change the
value of this parameter.

BUG=none
TEST=build ok, booted on daisy

Change-Id: I0ca3433aa78720b6dd1a8b451fa4ce7522905862
Signed-off-by: Hatim Ali <hatim.rv@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/22287
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>
Commit-Ready: Doug Anderson <dianders@chromium.org>