gen: Adjust fdt console to be silent if no alias present
At present, if there is no valid serial console alias, we hang early in
boot without a console.

Adjust this to continue with a silent console instead.

BUG=chrome-os-partner:7952
TEST=manual
build and boot on Seaboard
See that console does not appear when 'console' alias is removed from
fdt, but U-Boot still boots up on the LCD.

Change-Id: Id9aa7766e2be88f144f842e275306c7a4ad5be82
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/22277