exynos: i2c: Enable an early i2c register base address for early boot i2c access
This CL adds in an early i2c register base address, enabling functions in
early boot (SPL) to access a single I2C bus. To utilize this functionality
one would call set_early_i2c_reg from the early boot code for a particular
platform, passing it the base address of the peripheral. This effectively
allows you to replace the output of the samsung_get_base_ (i2c) function
from cpu.h with a custom value early on.

This is done in preparation for using FDT to configure i2c.

This currently modifies i2c.h as it is adding a new external API to the i2c
subsystem.

BUG=chromium-os:30663
TEST=Manual, verify 'i2c md 9 14 1' still gives 38.
Verify kernel still comes up properly.

Change-Id: Ia4a9b82c3d698f7adea58d67c8daa7afa0ba39ed
Signed-off-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/22314