drop: power: Make the max77686 driver chose an I2C bus explicitly
The max77686 driver uses I2C to communicate with the max77686 component, but
wasn't explicitly telling the I2C controller driver what bus to use,
implicitly selecting whichever one the controller driver happened to
initialize itself to. This turned out to be 0 which was correct, but if
something else runs before hand and changes the bus, this driver will blindly
start using that one instead.

This was happening when talking to the tps65090 driver to turn on power for
an LCDs backlight and panel. The max77686 driver attempted to read over the
wrong bus (4 in this case), that failed, and the bus was left in a bad state.
The bus was still that way when Linux tried to use it later, resulting in some
errors.

BUG=chrome-os-partner:9590
TEST=With this and another change, booted into the kernel and saw errors
related to i2c go away.

Change-Id: Ic7ff72a8a29a6da7fc1aee3e4b2a5426701299ec
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/22327
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>
Commit-Ready: Gabe Black <gabeblack@chromium.org>