exynos: fdt: Move SPI keyboard support from common to Daisy specific FDT
This CL moves the SPI keyboard node from the common dtsi file to the daisy
specific dts file, as not all CrOS devies based on 5250 will have a SPI
keyboard available.

A follow on to this will be to add an i2c node in the common dtsi file, as
all are expected to have i2c available.

BUG=chrome-os-partner:8869
TEST=Manual, verify fdt still builds properly.
Verify u-boot still boots and kernel comes up properly.

Change-Id: Ifebd5bf4884ec1906aee58e52d91f9de5d491d8e
Signed-off-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/22478
Reviewed-by: Simon Glass <sjg@chromium.org>